+++ RELEASE_NOTES.md
@@ -4,7 +4,8 @@
 
 CHANGES:
 
-1. Add val2sphp helper method.
+1. Add val2sphp helper method to convert high precision (9dp) coordinate to separate standard and high precision components, as required by some CFG and NAV messages.
+1. Add utc2itow helper method to convert utc datetime to GPS week number and time of week.
 
 ### RELEASE 1.2.37